Function abs() now accepts all types of NCL numeric data as input, instead of only integer data.  Its returned result is of the same type as the input.  abs() now supercedes fabs(), which accepts data of type float or double, and is still retained.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8963 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/AddIntrinsics.c
@@ -1,7 +1,7 @@
 
 
 /*
- *      $Id: AddIntrinsics.c,v 1.32 1999-04-23 22:07:48 ethan Exp $
+ *      $Id: AddIntrinsics.c,v 1.33 2006-06-07 21:52:40 grubin Exp $
  */
 /************************************************************************
 *									*
@@ -763,7 +763,7 @@ void _NclAddIntrinsics
 	_NclRegisterFunc(_NclIgetenv,args,"getenv",1,IFUNC);
 
 	args = NclCalloc(1,sizeof(NclArgTemplate));
-	args[0].arg_data_type = _NclLookUp("integer");
+	args[0].arg_data_type = _NclLookUp("numeric");
 	args[0].is_dimsizes = 0;
 	_NclRegisterFunc(_NclIabs,args,"abs",1,IFUNC);
